IBM_Blumix_DropBox_command_line/requirements.txt
Babel==2.3.4
certifi==2017.4.17
cf-deployment-tracker==1.0.4
chardet==3.0.3
click==6.7
cloudant==2.4.0
debtcollector==1.14.0
Flask==0.12.2
funcsigs==1.0.2
futures==3.1.1
idna==2.5
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.9.6
keystoneauth1==2.20.0
MarkupSafe==1.0
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.6
oslo.config==4.3.0
oslo.i18n==3.15.3
oslo.serialization==2.18.0
oslo.utils==3.26.0
pbr==3.0.1
positional==1.1.1
pycrypto==2.6.1
pyparsing==2.2.0
python-keystoneclient==3.10.0
python-swiftclient==3.3.0
pytz==2017.2
requests==2.17.3
rfc3986==1.0.0
simple-crypt==4.1.7
six==1.10.0
stevedore==1.23.0
urllib3==1.21.1
Werkzeug==0.12.2
wrapt==1.10.10
